+++ src/map/util.rs
@@ -1,38 +1,29 @@
 use super::*;
 
-// //二次元配列の添え字から画面座標を算出する
-// pub fn conv_sprite_coordinates( x: i32, y: i32 ) -> ( f32, f32 )
-// {	let x = ( PIXEL_PER_GRID - SCREEN_WIDTH  ) / 2.0 + PIXEL_PER_GRID * x as f32;
-// 	let y = ( SCREEN_HEIGHT - PIXEL_PER_GRID ) / 2.0 - PIXEL_PER_GRID * y as f32 - PIXEL_PER_GRID;
-// 	( x, y )
-// }
-
-////////////////////////////////////////////////////////////////////////////////////////////////////
-
 //壁判定のメソッド: is_wall()系 -> true: 壁である、false: 壁ではない
 impl GameMap
-{	pub fn is_wall( &self, x: i32, y: i32 ) -> bool
+{	pub fn is_wall( &self, x: usize, y: usize ) -> bool
 	{	if ! MAP_INDEX_X.contains( &x ) || ! MAP_INDEX_Y.contains( &y ) { return true } //配列の添字外は壁
 		matches!( self.map[ x as usize ][ y as usize ], MapObj::Wall(_) )
 	}
-	pub fn is_wall_upper_left   ( &self, x: i32, y: i32 ) -> bool { self.is_wall( x - 1, y - 1 ) }
-	pub fn is_wall_upper_center ( &self, x: i32, y: i32 ) -> bool { self.is_wall( x    , y - 1 ) }
-	pub fn is_wall_upper_right  ( &self, x: i32, y: i32 ) -> bool { self.is_wall( x + 1, y - 1 ) }
-	pub fn is_wall_middle_left  ( &self, x: i32, y: i32 ) -> bool { self.is_wall( x - 1, y     ) }
-	pub fn is_wall_middle_right ( &self, x: i32, y: i32 ) -> bool { self.is_wall( x + 1, y     ) }
-	pub fn is_wall_lower_left   ( &self, x: i32, y: i32 ) -> bool { self.is_wall( x - 1, y + 1 ) }
-	pub fn is_wall_lower_center ( &self, x: i32, y: i32 ) -> bool { self.is_wall( x    , y + 1 ) }
-	pub fn is_wall_lower_right  ( &self, x: i32, y: i32 ) -> bool { self.is_wall( x + 1, y + 1 ) }
+	pub fn is_wall_upper_left   ( &self, x: usize, y: usize ) -> bool { self.is_wall( x - 1, y - 1 ) }
+	pub fn is_wall_upper_center ( &self, x: usize, y: usize ) -> bool { self.is_wall( x    , y - 1 ) }
+	pub fn is_wall_upper_right  ( &self, x: usize, y: usize ) -> bool { self.is_wall( x + 1, y - 1 ) }
+	pub fn is_wall_middle_left  ( &self, x: usize, y: usize ) -> bool { self.is_wall( x - 1, y     ) }
+	pub fn is_wall_middle_right ( &self, x: usize, y: usize ) -> bool { self.is_wall( x + 1, y     ) }
+	pub fn is_wall_lower_left   ( &self, x: usize, y: usize ) -> bool { self.is_wall( x - 1, y + 1 ) }
+	pub fn is_wall_lower_center ( &self, x: usize, y: usize ) -> bool { self.is_wall( x    , y + 1 ) }
+	pub fn is_wall_lower_right  ( &self, x: usize, y: usize ) -> bool { self.is_wall( x + 1, y + 1 ) }
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
 //MAPの範囲の定数
 use std::ops::RangeInclusive;
-pub const MAP_INDEX_X  : RangeInclusive<i32> = 0..= MAP_WIDTH  - 1;	//MAP配列の添え字のレンジ
-pub const MAP_INDEX_Y  : RangeInclusive<i32> = 0..= MAP_HEIGHT - 1;	//MAP配列の添え字のレンジ
-pub const MAP_DIGABLE_X: RangeInclusive<i32> = 1..= MAP_WIDTH  - 2;	//掘削可能なレンジ（最外壁は掘れない）
-pub const MAP_DIGABLE_Y: RangeInclusive<i32> = 1..= MAP_HEIGHT - 2;	//掘削可能なレンジ（最外壁は掘れない）
+pub const MAP_INDEX_X  : RangeInclusive<usize> = 0..= MAP_WIDTH  - 1;	//MAP配列の添え字のレンジ
+pub const MAP_INDEX_Y  : RangeInclusive<usize> = 0..= MAP_HEIGHT - 1;	//MAP配列の添え字のレンジ
+pub const MAP_DIGABLE_X: RangeInclusive<usize> = 1..= MAP_WIDTH  - 2;	//掘削可能なレンジ（最外壁は掘れない）
+pub const MAP_DIGABLE_Y: RangeInclusive<usize> = 1..= MAP_HEIGHT - 2;	//掘削可能なレンジ（最外壁は掘れない）
 
 //MAP座標の上下左右を表す定数
 pub const UP   : ( i32, i32 ) = (  0, -1 );
@@ -53,22 +44,22 @@ const BIT_IS_EVENT_DONE: usize = 0b1000;
 impl GameMap
 {	//指定されたマスのフラグを返す
 	#[allow(dead_code)]
-	pub fn is_visible    ( &self, x: i32, y: i32 ) -> bool { self.stat[ x as usize ][ y as usize ] & BIT_IS_VISIBLE    != 0 }
-	pub fn is_passageway ( &self, x: i32, y: i32 ) -> bool { self.stat[ x as usize ][ y as usize ] & BIT_IS_PASSAGEWAY != 0 }
-	pub fn is_dead_end   ( &self, x: i32, y: i32 ) -> bool { self.stat[ x as usize ][ y as usize ] & BIT_IS_DEAD_END   != 0 }
-	pub fn is_event_done ( &self, x: i32, y: i32 ) -> bool { self.stat[ x as usize ][ y as usize ] & BIT_IS_EVENT_DONE != 0 }
+	pub fn is_visible    ( &self, x: usize, y: usize ) -> bool { self.stat[ x ][ y ] & BIT_IS_VISIBLE    != 0 }
+	pub fn is_passageway ( &self, x: usize, y: usize ) -> bool { self.stat[ x ][ y ] & BIT_IS_PASSAGEWAY != 0 }
+	pub fn is_dead_end   ( &self, x: usize, y: usize ) -> bool { self.stat[ x ][ y ] & BIT_IS_DEAD_END   != 0 }
+	pub fn is_event_done ( &self, x: usize, y: usize ) -> bool { self.stat[ x ][ y ] & BIT_IS_EVENT_DONE != 0 }
 
 	//指定されたマスのフラグを立てる
-	pub fn set_flag_passageway ( &mut self, x: i32, y: i32 ) { self.stat[ x as usize ][ y as usize ] |= BIT_IS_PASSAGEWAY; }
-	pub fn set_flag_dead_end   ( &mut self, x: i32, y: i32 ) { self.stat[ x as usize ][ y as usize ] |= BIT_IS_DEAD_END;   }
-	pub fn set_flag_event_done ( &mut self, x: i32, y: i32 ) { self.stat[ x as usize ][ y as usize ] |= BIT_IS_EVENT_DONE; }
+	pub fn set_flag_passageway ( &mut self, x: usize, y: usize ) { self.stat[ x ][ y ] |= BIT_IS_PASSAGEWAY; }
+	pub fn set_flag_dead_end   ( &mut self, x: usize, y: usize ) { self.stat[ x ][ y ] |= BIT_IS_DEAD_END;   }
+	pub fn set_flag_event_done ( &mut self, x: usize, y: usize ) { self.stat[ x ][ y ] |= BIT_IS_EVENT_DONE; }
 
 	//指定されたマスのVISIBLEフラグを立ててスプライトを可視化する
-	pub fn show( &mut self, x: i32, y: i32, q: &mut Query<&mut Visibility> )
+	pub fn show( &mut self, x: usize, y: usize, q: &mut Query<&mut Visibility> )
 	{	if ! MAP_INDEX_X.contains( &x ) || ! MAP_INDEX_Y.contains( &y ) { return }
 
-		self.stat[ x as usize ][ y as usize ] |= BIT_IS_VISIBLE;
-		if let MapObj::Wall( Some( id ) ) = self.map[ x as usize ][ y as usize ]
+		self.stat[ x ][ y ] |= BIT_IS_VISIBLE;
+		if let MapObj::Wall( Some( id ) ) = self.map[ x ][ y ]
 		{	q.get_component_mut::<Visibility>( id ).unwrap().is_visible = true;
 		}
 	}
